+++ front-end/src/components/DashboardNavBar.js
@@ -7,7 +7,7 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
-import DiaryMainView from './DiaryMainView';
+import DiaryMainView from './Diary/DiaryMainView';
 import Message from './Message';
 import Account from './Account';
 function TabPanel(props) {
