+++ front-end/src/components/Diary/DiaryMainView.css
