+++ front-end/src/components/Diary/DiaryMainView.js
@@ -27,8 +27,8 @@ import Tab from '@material-ui/core/Tab';
 import TabContext from '@material-ui/lab/TabContext';
 import TabList from '@material-ui/lab/TabList';
 import TabPanel from '@material-ui/lab/TabPanel';
-import Message from './Message';
-import Diary from './Diary';
+import Message from '../Message';
+import Diary from './diary/Diary';
 
 const useStyles = makeStyles((theme) => ({
   root: {
